+++ Sources/Schemas/CartesiaExperimentalControlsSpeedZero.swift
@@ -1,9 +1,0 @@
-import Foundation
-
-public enum CartesiaExperimentalControlsSpeedZero: String, Codable, Hashable, CaseIterable, Sendable {
-    case slowest
-    case slow
-    case normal
-    case fast
-    case fastest
-}