+++ Sources/Schemas/CartesiaSpeedControl.swift
@@ -1,13 +1,13 @@
 import Foundation
 
 public enum CartesiaSpeedControl: Codable, Hashable, Sendable {
-    case cartesiaExperimentalControlsSpeedZero(CartesiaExperimentalControlsSpeedZero)
+    case cartesiaSpeedControlZero(CartesiaSpeedControlZero)
     case double(Double)
 
     public init(from decoder: Decoder) throws {
         let container = try decoder.singleValueContainer()
-        if let value = try? container.decode(CartesiaExperimentalControlsSpeedZero.self) {
-            self = .cartesiaExperimentalControlsSpeedZero(value)
+        if let value = try? container.decode(CartesiaSpeedControlZero.self) {
+            self = .cartesiaSpeedControlZero(value)
         } else if let value = try? container.decode(Double.self) {
             self = .double(value)
         } else {
@@ -21,7 +21,7 @@ public enum CartesiaSpeedControl: Codable, Hashable, Sendable {
     public func encode(to encoder: Encoder) throws -> Void {
         var container = encoder.singleValueContainer()
         switch self {
-        case .cartesiaExperimentalControlsSpeedZero(let value):
+        case .cartesiaSpeedControlZero(let value):
             try container.encode(value)
         case .double(let value):
             try container.encode(value)