+++ Sources/Schemas/FallbackAzureVoiceId.swift
@@ -2,13 +2,13 @@ import Foundation
 
 /// This is the provider-specific ID that will be used.
 public enum FallbackAzureVoiceId: Codable, Hashable, Sendable {
-    case fallbackAzureVoiceVoiceIdZero(FallbackAzureVoiceVoiceIdZero)
+    case fallbackAzureVoiceIdZero(FallbackAzureVoiceIdZero)
     case string(String)
 
     public init(from decoder: Decoder) throws {
         let container = try decoder.singleValueContainer()
-        if let value = try? container.decode(FallbackAzureVoiceVoiceIdZero.self) {
-            self = .fallbackAzureVoiceVoiceIdZero(value)
+        if let value = try? container.decode(FallbackAzureVoiceIdZero.self) {
+            self = .fallbackAzureVoiceIdZero(value)
         } else if let value = try? container.decode(String.self) {
             self = .string(value)
         } else {
@@ -22,7 +22,7 @@ public enum FallbackAzureVoiceId: Codable, Hashable, Sendable {
     public func encode(to encoder: Encoder) throws -> Void {
         var container = encoder.singleValueContainer()
         switch self {
-        case .fallbackAzureVoiceVoiceIdZero(let value):
+        case .fallbackAzureVoiceIdZero(let value):
             try container.encode(value)
         case .string(let value):
             try container.encode(value)